Basic interface skeleton
Very messy code tbh, but is good as prototype. Need to study more about C++ and CMake. And need to separate the multiple Windows in classes in order to achive good modularity and, i expect, not fuck (entirely) with code. Probably will use dependency injection in main interface with all screens