Intial openFile() function (with nfd) and initial room validation
Still there's a lot to do (and i'll put more TODO's in code, it's better for organization). In fact, i need to discover *how* (and why) to stop nfd to stop my app when picking a file. Btw, delagate memory managment to compiler is so much better. Just use std::vector for example avoided a lot of bugs related to bad memory allocation.