Basic multi-window system
Created a basic multi-window system by multiple-context handling and multiple sdl objects creation. There's a very annoying input lag that i need to fix... I'm thinking on use threads for it. Have no f idea if this is a good idea or even if it will work, but Gott Mit Uns.