Changed main CMake, fixed some bugs, added Interface-Screen enum
Basically, changed main CMake to put the main project name as 'BABYLON'. Fixed a little bug in Controller where when a file is canceled with nfd the program crashes due to lack of validation if file is empty. Added at Interface class a enum to each screen and changed from std::vector to std::array (because it's a fixed size of possible screens)