Created basic fetch-decode cycle skeleton
I've added a lot of comments explaining what i shoud next. But to resume, i need to change the functions signatures to match (if is data so it recives a uint8_t and if is adresses recive a uint16_t). Need to implement properly the function that retrives the data on memory (based on addressing) too